requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn>=1.3.2
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
google-generativeai==0.3.2
geopy==2.4.1
werkzeug==3.0.1
gunicorn==21.2.0
